+++ Polygon_mesh_processing/test/Polygon_mesh_processing/test_stitching.cpp
@@ -52,7 +52,7 @@ void test_surface_mesh(const char* fname)
   CGAL::Polygon_mesh_processing::stitch_borders(m);
   //todo : add a validity test
 
-  assert(is_valid(m));
+  assert(is_valid_polygon_mesh(m));
 
   std::cout << "OK\n";
 }