+++ Polyhedron/demo/Polyhedron/include/id_printing.h
@@ -116,11 +116,11 @@ bool find_primitive_id(const QPoint& point,
 {
   typedef typename CGAL::Kernel_traits<Point>::Kernel Traits;
   bool found = false;
-  qglviewer::Vec point_under = viewer->camera()->pointUnderPixel(point,found);
-  const qglviewer::Vec offset = static_cast<CGAL::Three::Viewer_interface*>(QGLViewer::QGLViewerPool().first())->offset();
+  CGAL::qglviewer::Vec point_under = viewer->camera()->pointUnderPixel(point,found);
+  const CGAL::qglviewer::Vec offset = static_cast<CGAL::Three::Viewer_interface*>(CGAL::QGLViewer::QGLViewerPool().first())->offset();
 
   //find clicked facet
-  qglviewer::Vec dir = point_under - viewer->camera()->position();
+  CGAL::qglviewer::Vec dir = point_under - viewer->camera()->position();
   const Point ray_origin(viewer->camera()->position().x - offset.x,
                          viewer->camera()->position().y - offset.y,
                          viewer->camera()->position().z - offset.z);
@@ -169,7 +169,7 @@ void compute_displayed_ids(Mesh& mesh,
                            CGAL::Three::Viewer_interface *viewer,
                            const typename boost::graph_traits<Mesh>::face_descriptor& selected_fh,
                            const Point& pt_under,
-                           const qglviewer::Vec& offset,
+                           const CGAL::qglviewer::Vec& offset,
                            TextListItem* vitems,
                            TextListItem* eitems,
                            TextListItem* fitems,
@@ -390,7 +390,7 @@ bool printVertexIds(const Mesh& mesh,
   Ppmap ppmap = get(boost::vertex_point, mesh);
   IDmap idmap = get(boost::vertex_index, mesh);
   TextRenderer *renderer = viewer->textRenderer();
-  const qglviewer::Vec offset = viewer->offset();
+  const CGAL::qglviewer::Vec offset = viewer->offset();
   QFont font;
   font.setBold(true);
 
@@ -425,7 +425,7 @@ bool printEdgeIds(const Mesh& mesh,
   Ppmap ppmap = get(boost::vertex_point, mesh);
   IDmap idmap = get(boost::halfedge_index, mesh);
   TextRenderer *renderer = viewer->textRenderer();
-  const qglviewer::Vec offset = viewer->offset();
+  const CGAL::qglviewer::Vec offset = viewer->offset();
   QFont font;
   font.setBold(true);
 
@@ -458,7 +458,7 @@ bool printFaceIds(const Mesh& mesh,
   Ppmap ppmap = get(boost::vertex_point, mesh);
   IDmap idmap = get(boost::face_index, mesh);
   TextRenderer *renderer = viewer->textRenderer();
-  const qglviewer::Vec offset = viewer->offset();
+  const CGAL::qglviewer::Vec offset = viewer->offset();
   QFont font;
   font.setBold(true);
   BOOST_FOREACH(typename boost::graph_traits<Mesh>::face_descriptor fh, faces(mesh))
@@ -518,7 +518,7 @@ int zoomToId(const Mesh& mesh,
   {
     return 1; //("Input must be of the form [v/e/f][int]"
   }
-  const qglviewer::Vec offset = viewer->offset();
+  const CGAL::qglviewer::Vec offset = viewer->offset();
   typename Traits::Vector_3 normal;
   if(first == QString("v"))
   {
@@ -599,24 +599,18 @@ int zoomToId(const Mesh& mesh,
       return 4; //"No face with id %1").arg(id)
     }
   }
-  qglviewer::Quaternion new_orientation(qglviewer::Vec(0,0,-1),
-                                        qglviewer::Vec(-normal.x(), -normal.y(), -normal.z()));
+  CGAL::qglviewer::Quaternion new_orientation(CGAL::qglviewer::Vec(0,0,-1),
+                                        CGAL::qglviewer::Vec(-normal.x(), -normal.y(), -normal.z()));
   Point new_pos = p +
-      qglviewer::Vec(
+      CGAL::qglviewer::Vec(
         viewer->camera()->position().x - viewer->camera()->sceneCenter().x,
         viewer->camera()->position().y - viewer->camera()->sceneCenter().y,
         viewer->camera()->position().z - viewer->camera()->sceneCenter().z)
       .norm() * normal ;
 
-#if QGLVIEWER_VERSION >= 0x020502
-  viewer->camera()->setPivotPoint(qglviewer::Vec(p.x(),
+  viewer->camera()->setPivotPoint(CGAL::qglviewer::Vec(p.x(),
                                                  p.y(),
                                                  p.z()));
-#else
-  viewer->camera()->setRevolveAroundPoint(qglviewer::Vec(p.x(),
-                                                         p.y(),
-                                                         p.z()));
-#endif
 
   viewer->moveCameraToCoordinates(QString("%1 %2 %3 %4 %5 %6 %7").arg(new_pos.x())
                                   .arg(new_pos.y())