+++ Polyhedron/include/CGAL/boost/graph/graph_traits_Polyhedron_3.h
@@ -464,12 +464,6 @@ reserve(CGAL::Polyhedron_3<Gt,I,HDS,A>& p,
 }
 
 template<class Gt, class I, CGAL_HDS_PARAM_, class A>
-bool is_valid(const CGAL::Polyhedron_3<Gt,I,HDS,A>& p, bool verbose = false)
-{
-  return p.is_valid(verbose);
-}
-
-template<class Gt, class I, CGAL_HDS_PARAM_, class A>
 void normalize_border(CGAL::Polyhedron_3<Gt,I,HDS,A>& p)
 {
   p.normalize_border();